+++ changelog.md
@@ -1,3 +1,4 @@
 # Version 3
 
 - Removal of setOption method on the control
+- getTile is now getBlobByKey